Plugin architecture for parsers
